requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
beautifulsoup4
google-genai
pyparsing
dotenv
ddgs
python-multipart
supabase
torch 
torchvision 
opencv-python 
numpy
apscheduler
pydantic